Adding Loging System

+++ include/clsMainScreen.h
@@ -11,6 +11,7 @@
 #include "clsDeleteClientScreen.h"
 #include "clsTransactionScreen.h" 
 #include "clsManageUsersScreen.h"
+#include "clsLoginScreen.h"
 
 class clsMainScreen : protected clsScreen
 {
@@ -36,7 +37,6 @@ private:
     {
         clsScreen::Print("\n\n");
         clsScreen::_PauseScreen();
-        ShowMainMenue();
     } 
     static void _ClientsListScreen()
     {
@@ -77,32 +77,80 @@ private:
     static void _LoginScreen()
     {
         system("cls");
-        clsScreen::Print("Here will be login screen\n");
+        clsLoginScreen::ShowLoginScreen();
     }
 
-    static void _PerformMainMenueChoice(enMainMenueOption choice)
+    static void _PerformMainMenueChoice(enMainMenueOption choice, clsUser::enUserPermissions UserPermissions)
     {
         switch (choice)
         {
         case eShowClientsList:
+            if (!(UserPermissions & clsUser::enUserPermissions::showClientList))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _ClientsListScreen();
             break;
+
         case eFindClient:
+            if (!(UserPermissions & clsUser::enUserPermissions::findClient))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _FindCliendScreen();
             break;
+
         case eAddNewClient:
+            if (!(UserPermissions & clsUser::enUserPermissions::addClient))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _AddNewClientScreen();
             break;
+
         case eUpdateClientData:
+            if (!(UserPermissions & clsUser::enUserPermissions::updateClient))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _UpdateClientScreen();
             break;
+
         case eDeleteClient:
+            if (!(UserPermissions & clsUser::enUserPermissions::deleteClient))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _DeleteClientScreen();
             break;
+
         case eTransactionsMenue:
+            if (!(UserPermissions & clsUser::enUserPermissions::transactions))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _TransactionsScreen();
             break;
+
         case eManageUsersMenue:
+            if (!(UserPermissions & clsUser::enUserPermissions::manageUsers))
+            {
+                clsScreen::Print("\n");
+                clsScreen::Print("<<< You Don't Have Permission To Do That. >>>");
+                break;
+            }
             _ManageUsersScreen();           
             break;
         case eLogout:
@@ -119,22 +167,26 @@ private:
     }
 
 public:
-    static void ShowMainMenue()
+    static void ShowMainMenue(clsUser::enUserPermissions UserPermissions)
     {
-        system("cls");
-        _PrintHeader("Main Screen", 5);
-        clsScreen::Print("=================================\n");
-        clsScreen::Print(" [1] Show Clients List.\n");
-        clsScreen::Print(" [2] Find Client.\n");
-        clsScreen::Print(" [3] Add New Client.\n");
-        clsScreen::Print(" [4] Update Client Data.\n");
-        clsScreen::Print(" [5] Delete Client.\n");
-        clsScreen::Print(" [6] Transactions Menue.\n");
-        clsScreen::Print(" [7] Manage Useres Menue\n");
-        clsScreen::Print(" [8] Logout.\n");
-        clsScreen::Print("=================================\n\n");
-        
-        enMainMenueOption choice =  _ReadMainMenueOption();
-        _PerformMainMenueChoice(choice);
+        while (true)
+        {
+            system("cls");
+            _PrintHeader("Main Screen", 5);
+            clsScreen::Print("=================================\n");
+            clsScreen::Print(" [1] Show Clients List.\n");
+            clsScreen::Print(" [2] Find Client.\n");
+            clsScreen::Print(" [3] Add New Client.\n");
+            clsScreen::Print(" [4] Update Client Data.\n");
+            clsScreen::Print(" [5] Delete Client.\n");
+            clsScreen::Print(" [6] Transactions Menue.\n");
+            clsScreen::Print(" [7] Manage Useres Menue\n");
+            clsScreen::Print(" [8] Logout.\n");
+            clsScreen::Print("=================================\n\n");
+
+            enMainMenueOption choice = _ReadMainMenueOption();
+            _PerformMainMenueChoice(choice, clsLoginScreen::CurrentUserPermissions);
+        }
+
     }
 };
